+++ packages/project-input-validator/src/index.ts
@@ -2,3 +2,5 @@ export * from './deploy/validators';
 export * from './ui5/validators';
 export * from './general/validators';
 export * from './adp/validators';
+export * from './flp/validators';
+export { addi18nResourceBundle } from './i18n';