+++ packages/abap-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/abap-deploy-config-inquirer
 
+## 1.1.0
+
+### Minor Changes
+
+-   849ef1e: updates options and conditions for prompts
+
+## 1.0.26
+
+### Patch Changes
+
+-   aa0731f: ensure scp prompt is maintained
+
+## 1.0.25
+
+### Patch Changes
+
+-   Updated dependencies [74dc5fe]
+    -   @sap-ux/inquirer-common@0.5.4
+
 ## 1.0.24
 
 ### Patch Changes