+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sap-ux/adp-tooling
 
+## 0.12.85
+
+### Patch Changes
+
+-   Updated dependencies [cfdd442]
+    -   @sap-ux/project-input-validator@0.3.4
+
+## 0.12.84
+
+### Patch Changes
+
+-   6e60bdf: Store manifest change files to /changes directory
+
+## 0.12.83
+
+### Patch Changes
+
+-   09a58bb: chore: upgrade vocabularies-types + pnpm updates
+
+## 0.12.82
+
+### Patch Changes
+
+-   Updated dependencies [74dc5fe]
+    -   @sap-ux/inquirer-common@0.5.4
+
 ## 0.12.81
 
 ### Patch Changes