+++ packages/preview-middleware-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux-private/preview-middleware-client
 
+## 0.11.36
+
+### Patch Changes
+
+-   71bef63: fix: update quick action title for semantic date range
+
+## 0.11.35
+
+### Patch Changes
+
+-   c39325c: Fix for Config Quick action state not reflecting properties in properties panel
+
 ## 0.11.34
 
 ### Patch Changes